move code into a separate file
